backend/pyproject.toml
[tool.poetry]
name = "bio-data-harmoniser"
version = "0.1.0"
description = ""
authors = ["BenTenmann <benji.tenmann@me.com>"]
readme = "README.md"
packages = [{include = "bio_data_harmoniser"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
streamlit = "^1.31.0"
openai = "^1.12.0"
deltalake = "^0.15.3"
loguru = "^0.7.2"
streamlit-code-editor = "^0.1.10"
pandasql = "^0.7.3"
faiss-cpu = "^1.7.4"
lxml = "^5.1.0"
networkx = "^3.2.1"
pyvis = "^0.3.2"
cloudpathlib = {extras = ["gs"], version = "^0.17.0"}
watchdog = "^4.0.0"
extra-streamlit-components = "^0.1.70"
streamlit-mermaid = "^0.2.0"
streamlit-lottie = "^0.0.5"
sentence-transformers = "^2.5.1"
torch = [
    { url = "https://download.pytorch.org/whl/cpu/torch-2.1.1%2Bcpu-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'"},
    { url = "https://download.pytorch.org/whl/cpu/torch-2.1.1-cp311-none-macosx_10_9_x86_64.whl", markers = "sys_platform == 'darwin' and platform_machine != 'arm64'"},
    { url = "https://download.pytorch.org/whl/cpu/torch-2.1.1-cp311-none-macosx_11_0_arm64.whl", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
]
# torch = "^2.2.1"  # for gpu
oxbow = "^0.3.1"
polars = "^0.20.16"
ieugwaspy = "^0.1.11"
openpyxl = "^3.1.2"
urlextract = "^1.9.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
apache-airflow = {extras = ["celery", "postgres"], version = "^2.9.0"}
redis = "^5.0.3"
psycopg2 = "^2.9.9"
pandera = {extras = ["io"], version = "^0.19.3"}
pydantic-settings = "^2.2.1"
html2text = "^2024.2.26"
playwright = "^1.43.0"
beautifulsoup4 = "^4.12.3"
rdkit = "^2023.9.6"
sqlglot = "^24.1.0"
langchain-openai = "^0.1.8"
langchain-anthropic = "^0.1.15"
langchain-text-splitters = "^0.2.1"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"
owlready2 = "^0.45"
typer = "^0.9.0"
pytest = "^8.0.1"
ipykernel = "^6.29.2"
jupyterlab = "^4.1.2"
black = {extras = ["jupyter"], version = "^24.2.0"}
isort = "^5.13.2"
ipywidgets = "^8.1.2"
line-profiler = "^4.1.2"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"

[tool.poetry.extras]
chem = ["rdkit"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bio-data-harmoniser = "bio_data_harmoniser.api.cli:main"
